Refactor lalloc additions.
