+++ docs/checklist.md
@@ -33,15 +33,23 @@
 - [ ] Not use more than one global variable. Think about it. You will have to explain its purpose.
 - [ ] Pass all testers
   - [ ] https://github.com/LucasKuhn/myshell_tester
+- [ ] Add `debug()` messages:
+  - [ ] Prompt
+  - [ ] Expander
+  - [x] Lexer
+  - [ ] Parser
+  - [ ] Syntax Validator
+  - [ ] Trimmer
+  - [ ] Executor
 
 ## Memória
 
-- [ ] Adicionar toda memoria do loop (REPL) no `lalloc()`
-  - [ ] Prompt
-  - [ ] Expander
-  - [ ] Lexer
+- [x] Adicionar toda memoria do loop (REPL) no `lalloc()`
+  - [x] Prompt
+  - [x] Expander
+  - [x] Lexer
   - [x] Parser
-- [ ] Liberar memória antes de sair do loop com `free_lalloc()`
+- [x] Liberar memória antes de sair do loop com `free_lalloc()`
 - [x] Remover todos os usos de `destroy_parse()`
 
 ## Variáveis
@@ -168,28 +176,30 @@
 
 ## Syntax Validator
 
-- [ ] Aspas:
-  - [ ] Aspas simples que não fecham.
-  - [ ] Aspas duplas que não fecham.
-- [ ] Variáveis:
-  - [ ] Nome de variáveis não pode começar com número.
-  - [ ] Nome de variáveis não pode conter metacharacters:
+- [x] Aspas:
+  - [x] Aspas simples que não fecham.
+  - [x] Aspas duplas que não fecham.
+- [x] Variáveis:
+  - [x] Nome de variáveis não pode começar com número.
+  - [x] Nome de variáveis não pode conter metacharacters:
     - `?`, `'`, `"`, `\`, `$`, <code>\`</code>, `=`, `*`, `@`, `~`, `<`, `>`, `(`, `)`, `!`, `|`, `&`, `;`, `space`, `newline`
-- [ ] Nome de arquivo:
-  - [ ] Linux: não pode conter `/`.
-  - [ ] Não pode conter characters especias fora de aspas: `|`, `;`, <code>\`</code>, `(`, `)`, `!`, `!`.
-- [ ] Input do usuário:
-  - [ ] Não pode ter `;` fora de aspas.
-  - [ ] Não pode ter `\` fora de aspas.
-  - [ ] Não pode ter `>` sem nome do arquivo.
-  - [ ] Não pode ter `>>` sem nome do arquivo.
-  - [ ] Não pode ter `<` sem nome do arquivo.
-  - [ ] Não pode ter `<<` sem delimitador.
-  - [ ] Não pode ter `<<` seguido de outro operador: `|`, `>`, `>>`, `<`, `<<`.
-  - [ ] Não pode ter `|` sem próximo commando.
-  - [ ] Não pode ter `|` seguido de `|`.
-  - [ ] Primeiro token não pode ser `|`.
-  - [ ] Printa mensagem de erro.
+- [x] Nome de arquivo:
+  - [x] Linux: não pode conter `/`.
+  - [x] Não pode conter characters especias fora de aspas: `|`, `;`, <code>\`</code>, `(`, `)`, `!`, `!`.
+- [x] Input do usuário:
+  - [x] Não pode ter `;` fora de aspas.
+  - [x] Não pode ter `\` fora de aspas.
+  - [x] Não pode ter `>` sem nome do arquivo.
+  - [x] Não pode ter `>>` sem nome do arquivo.
+  - [x] Não pode ter `<` sem nome do arquivo.
+  - [x] Não pode ter `<<` sem delimitador.
+  - [x] Não pode ter `<<` seguido de outro operador: `|`, `>`, `>>`, `<`, `<<`.
+  - [x] Não pode ter `|` sem próximo commando.
+  - [x] Não pode ter `|` seguido de `|`.
+  - [x] Primeiro token não pode ser `|`.
+  - [ ] Printar mensagem de erro.
+- [x] Validar sintaxe dos `char **tokens`
+- [ ] (OPCIONAL) Validar sintaxe da `t_parse_list plist`
 
 ## Trimmer
