+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/21 11:42:09 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/10/07 15:45:25 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/07 21:06:18 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,8 +77,9 @@ char			**tokenize_variable(char *declaration);
  * EXPANDER
 \******************************************************************************/
 
-char			*expand(char *input);
+char			*expand(char *line);
 
+char			*expand_line(char *line);
 char			**isolate_variables(char *input);
 
 /******************************************************************************\
