+++ main.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/25 10:34:20 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/10/07 16:23:52 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/07 21:15:52 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,12 +14,7 @@
 
 static void	initialize(int argc, char **argv, char **envp)
 {
-	char	**env_vars;
-
 	initialize_shell(argc, argv, envp);
-	env_vars = envht_to_envp();
-	ft_put_strarr(env_vars);
-	ft_free_strarr(env_vars);
 	ft_clear();
 }
