+++ sources/mandatory/executor/execute.c
@@ -6,15 +6,24 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/25 10:39:54 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/09/28 12:00:22 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/07 21:13:24 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minishell.h>
 
-void	execute(t_parse_list *pipeline)
+static void	handle_pipeline(t_parse_list *pipeline)
 {
 	if (handled_single_builtin(pipeline))
 		return ;
 	execute_pipeline(pipeline);
+}
+
+void	execute(t_parse_list *pipeline)
+{
+	t_proc_fds	ioe;
+
+	save_ioe(&ioe);
+	handle_pipeline(pipeline);
+	restore_ioe(&ioe);
 }
