+++ sources/mandatory/executor/pipeline.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/13 15:33:02 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/10/07 17:04:10 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/07 17:46:01 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,9 +15,7 @@
 void	execute_pipeline(t_parse_list *pipeline)
 {
 	int			pipe_count;
-	t_proc_fds	ioe;
 
-	save_ioe(&ioe);
 	pipe_count = count_parse_pipes(pipeline);
 	while (pipe_count > 0)
 	{
@@ -26,5 +24,4 @@ void	execute_pipeline(t_parse_list *pipeline)
 		pipe_count--;
 	}
 	execute_last_pipe(pipeline);
-	restore_ioe(&ioe);
 }
