+++ sources/mandatory/prompt/get_line.c
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/31 11:53:14 by wwallas-          #+#    #+#             */
-/*   Updated: 2022/10/07 17:05:07 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/10/07 21:53:32 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,16 +30,19 @@ static char	*read_prompt(void)
 	char	*ps1;
 
 	ps1 = build_ps1();
-	line = readline(ps1);
+	line = NULL;
+	while (line == NULL)
+		line = readline(ps1);
 	free(ps1);
 	return (line);
 }
 
 char	*prompt(void)
 {
-	char	*cmd_line;
+	char	*line;
 
-	cmd_line = read_prompt();
-	add_history(cmd_line);
-	return (cmd_line);
+	line = read_prompt();
+	add_history(line);
+	ft_add_lalloc(lalloc(), line);
+	return (line);
 }
